Population Attributable Fraction

Estimates the population attributable fraction, paf, for individual-level exposure data (and covariates), X, from a cross-sectional survey. Exposure is assumed to be associated with a relative risk function, rr, with parameter theta. A counterfactual scenario as a function of the exposure cft is assumed.

The population attributable fraction is defined Chan et al. (2023) as: \[ \text{PAF} = \dfrac{\mathbb{E}\Big[RR(X;\theta)\Big] - 1}{\mathbb{E}\Big[RR(X;\theta)\Big]} \]

where:

\(X\) denotes the individual-level matrix of exposure and covariates,

\(\theta\) represents additional parameters of the relative risk function,

\(RR(X,\theta)\) denotes the relative risk of exposure (and covariates) at level \(X\) given parameters \(\theta\),

\({\mathbb{E}\Big[RR(X;\theta)\Big]}\) represents the average relative risk in the population

\(1\) represents the relative risk under the theoretical minimum risk scenario,

\(\text{PAF}\) represents the population attributable fraction

Usage

paf(
  design,
  theta,
  rr,
  additional_theta_arguments,
  n_bootstrap_samples = NULL,
  theta_distribution = "default",
  parallel = TRUE,
  num_cores = 1,
  weights = NULL,
  ...
)

Arguments

design

(survey.design, data.frame,tibble, or svyrep.design) survey data structure. If data comes from a survey set the design with survey::svydesign(). It can also support a survey::svrepdesign() design if your survey comes with replicates. Finally, the model can also accommodate a data.frame or tibble with weights assuming simple random sampling without replacement.

theta

(vector/double) parameters of the relative risk function rr.

rr

(function/list) a relative risk function with two parameters: a data.frame called X containing the individual-level exposure and covariates, and theta (in that order). It can also be a list of several relative risk functions to apply with each function being a different modelling scenario.

additional_theta_arguments

any additional information on theta utilized for obtaining bootstrap samples from the paramter. Options are:

(double) the variance of theta if theta is one dimensional and asymptotical normality is assumed (default).

(vector) the variances of each entry of theta if theta is n-dimensional and its entries are uncorrelated and asymptotical normality is assumed (default).

(matrix) the variance-covariance matrix of theta if theta is n-dimensional and its entries are correlated and asymptotical normality is assumed (default).

any list of arguments to pass via base::do.call() to theta_distribution to simulate samples from theta if theta is not assumed to be asymptotically normally distributed.

Optional

n_bootstrap_samples

(double) number of bootstrap samples. If a svyrep.design is passed as an argument, then n_bootstrap_samples represents the number of number of replicates in the design.

theta_distribution

(function) random number generator that follows the distribution of the estimator theta. By default, theta is assumed to be asymptotically normal and thus theta_distribution is set to mvtnorm::rmvnorm() with variance given by additional_theta_arguments. The number of simulations for the theta_distribution function must be parametrized by a parameter of name n.

parallel

(boolean) Flag indicating whether or not to do computations in parallel. Default is TRUE.

num_cores

(int) Number of cores; defaults to parallel::detectCores().

weights

(vector) If you are not following the recommended version and use a svydesign object for the design you can still use weights to associate weights to your estimation. Beware that it might not give accurate estimations of the variance nor the uncertainty intervals.

...

Additional parameters for svrep::as_bootstrap_design().

Confidence Intervals

Confidence intervals are estimated via the two methods specified by Arnab (2017) .

Wald-type confidence intervals (which are more precise) are of the form: \[ \widehat{\text{PAF}} \pm t_{1 - \alpha/2} \sqrt{\widehat{\text{Var}}_{\text{B}}\big[\widehat{\text{PIF}}\big]} \] where \(t_{1 - \alpha/2}\) stands for the percent points at level \(1 - \alpha/2\) of Student's t-distribution, and \(\widehat{\text{Var}}_{\text{B}}\big[\widehat{\text{PIF}}\big]\) represents the estimated variance (via bootstrap) of the potential impact fraction estimator.

Percentile confidence intervals (less precise) are of the form: \[ \Big[\widehat{\text{PAF}}_{\text{B},\alpha/2}, \widehat{\text{PAF}}_{\text{B},1-\alpha/2}\Big] \] where \(\widehat{\text{PAF}}_{\text{B},k}\) represents the kth sample quantile from the bootstrap simulation of the potential impact fraction estimator.

Additional parallelization options

By default the function uses foreach::foreach to parallelize creating a cluster via doParallel::registerDoParallel. For finner parallelization control you can do:

#Create the cluster outside using whatever you want
myCluster <- parallel::makeCluster(3, type = "PSOCK")

#Register the cluster
doParallel::registerDoParallel(myCluster)

#Call the function using parallel = TRUE to use %dopar% and num_cores = NULL to avoid setting
paf(..., parallel = TRUE, num_cores = NULL)

#Stop the cluster
doParallel::stopCluster(myCluster)

References

Arnab R (2017). “Survey sampling theory and applications.” In chapter Variance Estimation: Complex Survey Designs. Academic Press.

Chan CE, Zepeda-Tello R, Camacho-García-Formentí D, Cudhea F, Meza R, Rodrigues E, Spiegelman D, Barrientos-Gutierrez T, Zhou X (2023). “Nonparametric Estimation of the Potential Impact Fraction and Population Attributable Fraction with Individual-Level and Aggregated Data.” 2207.03597.

See also

Examples

# Use the ensanut dataset
data(ensanut)

# EXAMPLE 1
# Setup the survey design
options(survey.lonely.psu = "adjust")
design <- survey::svydesign(data = ensanut, ids = ~1, weights = ~weight, strata = ~strata)
rr <- function(X, theta) {
  exp(
    theta[1] * X[, "age"] + theta[2] * X[, "systolic_blood_pressure"] / 100)
}
paf(design,
  theta = log(c(1.05, 1.38)), rr,
  additional_theta_arguments = c(0.01, 0.03), n_bootstrap_samples = 10, parallel = FALSE
)
#> ── Population Attributable Fraction (PAF) ──────────────────────────────────────
#>   population_attributable_fraction           type average_relative_risk
#> 1                        0.2447751 point_estimate              18039863
#> 2                       -3.2920090     Lower 2.5%            -107919525
#> 3                        3.7815592    Upper 97.5%             143999251
#>   average_counterfactual   relative_risk                 counterfactual
#> 1                      1 Relative_Risk_1 Theoretical_minimum_risk_level
#> 2                      1 Relative_Risk_1 Theoretical_minimum_risk_level
#> 3                      1 Relative_Risk_1 Theoretical_minimum_risk_level
#> ────────────────────────────────────────────────────────────────────────────────
#> • Number of bootstrap simulations: 10
#> ✖ A low number of bootstrap simulations will result in an unstable estimate.
#> • Use `as.data.frame` to access values.
#> • Use `summary` to save list of main results.

# EXAMPLE 2
# Now do the same but using a replicate design
options(survey.lonely.psu = "adjust")
rep_design <- svrep::as_bootstrap_design(design, replicates = 10)
paf(rep_design,
  theta = log(c(1.05, 1.38)), rr,
  additional_theta_arguments = c(0.01, 0.03)
)
#> ── Population Attributable Fraction (PAF) ──────────────────────────────────────
#>   population_attributable_fraction           type average_relative_risk
#> 1                         -7.06308 point_estimate              47003.54
#> 2                        -53.79660     Lower 2.5%            -283313.88
#> 3                         39.67044    Upper 97.5%             377320.97
#>   average_counterfactual   relative_risk                 counterfactual
#> 1                      1 Relative_Risk_1 Theoretical_minimum_risk_level
#> 2                      1 Relative_Risk_1 Theoretical_minimum_risk_level
#> 3                      1 Relative_Risk_1 Theoretical_minimum_risk_level
#> ────────────────────────────────────────────────────────────────────────────────
#> • Number of bootstrap simulations: 10
#> ✖ A low number of bootstrap simulations will result in an unstable estimate.
#> • Use `as.data.frame` to access values.
#> • Use `summary` to save list of main results.