Potential Impact Fraction via bootstrap for survey data

Estimates the potential impact fraction, pif, for individual-level exposure data (and covariates), X, from a cross-sectional survey. Exposure is assumed to be associated with a relative risk function, rr, with parameter theta. A counterfactual scenario as a function of the exposure cft is assumed.

The potential impact fraction is defined Chan et al. (2023) : \[ \text{PIF} = \dfrac{\mathbb{E}\Big[RR(X;\theta)\Big] - \mathbb{E}\Big[RR\big(\text{cft}(X);\theta\big)\Big]}{\mathbb{E}\Big[RR(X;\theta)\Big]} \]

where:

\(X\) denotes the individual-level matrix of exposure and covariates,

\(\theta\) represents additional parameters of the relative risk function,

\(RR(X,\theta)\) denotes the relative risk of exposure (and covariates) at level \(X\) given parameters \(\theta\),

\(cft(X)\) denotes the counterfactual function applied to exposure and covariates,

\({\mathbb{E}\Big[RR(X;\theta)\Big]}\) and \({\mathbb{E}\Big[RR(\text{cft}(X);\theta)\Big]}\) denote the population average relative risk under current (observed) conditions and the relative risk under the counterfactual scenario.

\(\text{PIF}\) represents the potential impact fraction.

pif_survey_bootstrap()

Arguments

design

(survey.design/svyrep.design) survey design structure from the survey package obtained from survey::svydesign() or a survey replicates design obtained either from survey::svrepdesign() or svrep::as_bootstrap_design(). Contains individual level exposure and covariates.

theta

(vector/double) parameters of the relative risk function rr.

rr

(function) relative risk function with two parameters: X and theta (in that order).

cft

(function) counterfactual function with one parameter: X.

n_samples

(double) number of bootstrap samples

additional_theta_arguments

any additional information on theta utilized for obtaining bootstrap samples from the paramter. Options are:

(double) the variance of theta if theta is one dimensional and asymptotical normality is assumed (default).

(vector) the variances of each entry of theta if theta is n-dimensional and its entries are uncorrelated and asymptotical normality is assumed (default).

(matrix) the variance-covariance matrix of theta if theta is n-dimensional and its entries are correlated and asymptotical normality is assumed (default).

any argument to pass to theta_distribution to simulate samples from theta if theta is not assumed to be asymptotically normally distributed.

Optional

theta_distribution

(function) random number generator that follows the distribution of the estimator theta. By default, theta is assumed to be asymptotically normal and thus theta_distribution is set to mvtnorm::rmvnorm() with variance given by additional_theta_arguments.

uncertainty_interval_type

(string) either 'wald' (recommended) or 'percentile' for uncertainty intervals based upon the bootstrap's percentile or the Wald-type approximation (see confidence interval section)

parallel

(boolean) Flag indicating whether or not to do computations in parallel. Default is TRUE.

num_cores

(int) Number of cores; defaults to parallel::detectCores().

confidence_level

(double) Confidence level for the uncertainty interval. Defaults to 0.95

Confidence Intervals

Confidence intervals are estimated via the two methods specified by Arnab (2017) .

Wald-type confidence intervals (which are more precise) are of the form: \[ \widehat{\text{PIF}} \pm t_{1 - \alpha/2} \sqrt{\widehat{\text{Var}}_{\text{B}}\big[\widehat{\text{PIF}}\big]} \] where \(t_{1 - \alpha/2}\) stands for the percent points at level \(1 - \alpha/2\) of Student's t-distribution, and \(\widehat{\text{Var}}_{\text{B}}\big[\widehat{\text{PIF}}\big]\) represents the estimated variance (via bootstrap) of the potential impact fraction estimator.

Percentile confidence intervals (less precise) are of the form: \[ \Big[\widehat{\text{PIF}}_{\text{B},\alpha/2}, \widehat{\text{PIF}}_{\text{B},1-\alpha/2}\Big] \] where \(\widehat{\text{PIF}}_{\text{B},k}\) represents the kth sample quantile from the bootstrap simulation of the potential impact fraction estimator.

References

Arnab R (2017). “Survey sampling theory and applications.” In chapter Variance Estimation: Complex Survey Designs. Academic Press.

Chan CE, Zepeda-Tello R, Camacho-García-Formentí D, Cudhea F, Meza R, Rodrigues E, Spiegelman D, Barrientos-Gutierrez T, Zhou X (2023). “Nonparametric Estimation of the Potential Impact Fraction and Population Attributable Fraction with Individual-Level and Aggregated Data.” 2207.03597.

See also