Potential Impact Fraction

Estimates the potential impact fraction, pif, for individual-level exposure data (and covariates), X, from a cross-sectional survey. Exposure is assumed to be associated with a relative risk function, rr, with parameter theta. A counterfactual scenario as a function of the exposure cft is assumed.

The potential impact fraction is defined Chan et al. (2023) : \[ \text{PIF} = \dfrac{\mathbb{E}\Big[RR(X;\theta)\Big] - \mathbb{E}\Big[RR\big(\text{cft}(X);\theta\big)\Big]}{\mathbb{E}\Big[RR(X;\theta)\Big]} \]

where:

\(X\) denotes the individual-level matrix of exposure and covariates,

\(\theta\) represents additional parameters of the relative risk function,

\(RR(X,\theta)\) denotes the relative risk of exposure (and covariates) at level \(X\) given parameters \(\theta\),

\(cft(X)\) denotes the counterfactual function applied to exposure and covariates,

\({\mathbb{E}\Big[RR(X;\theta)\Big]}\) and \({\mathbb{E}\Big[RR(\text{cft}(X);\theta)\Big]}\) denote the population average relative risk under current (observed) conditions and the relative risk under the counterfactual scenario.

\(\text{PIF}\) represents the potential impact fraction.

Usage

pif(
  design,
  theta,
  rr,
  cft = NULL,
  additional_theta_arguments,
  n_bootstrap_samples = NULL,
  theta_distribution = "default",
  is_paf = FALSE,
  weights = NULL,
  .options.future = list(seed = TRUE),
  ...
)

Arguments

design

(survey.design, data.frame,tibble, or svyrep.design) survey data structure. If data comes from a survey set the design with survey::svydesign(). It can also support a survey::svrepdesign() design if your survey comes with replicates. Finally, the model can also accommodate a data.frame or tibble with weights assuming simple random sampling without replacement.

theta

(vector/double) parameters of the relative risk function rr.

rr

(function/list) a relative risk function with two parameters: a data.frame called X containing the individual-level exposure and covariates, and theta (in that order). It can also be a list of several relative risk functions to apply with each function being a different modelling scenario.

cft

(function/list) a counterfactual function that takes a data.frame, X, of individual-level exposure and covariates and returns a new data.frame of individual-level counterfactual exposure and covariates. It can also be a list of several counterfactual functions to apply with each function being a different modelling scenario.

additional_theta_arguments

any additional information on theta utilized for obtaining bootstrap samples from the paramter. Options are:

(double) the variance of theta if theta is one dimensional and asymptotical normality is assumed (default).

(vector) the variances of each entry of theta if theta is n-dimensional and its entries are uncorrelated and asymptotical normality is assumed (default).

(matrix) the variance-covariance matrix of theta if theta is n-dimensional and its entries are correlated and asymptotical normality is assumed (default).

any list of arguments to pass via base::do.call() to theta_distribution to simulate samples from theta if theta is not assumed to be asymptotically normally distributed.

Optional

n_bootstrap_samples

(double) number of bootstrap samples. If a svyrep.design is passed as an argument, then n_bootstrap_samples represents the number of number of replicates in the design.

theta_distribution

(function) random number generator that follows the distribution of the estimator theta. By default, theta is assumed to be asymptotically normal and thus theta_distribution is set to mvtnorm::rmvnorm() with variance given by additional_theta_arguments. The number of simulations for the theta_distribution function must be parametrized by a parameter of name n.

is_paf

(boolean) Whether the function being estimated is the Population Attributable Fraction (is_paf = TRUE) or the Potential Impact Fraction (is_paf = FALSE)

weights

(vector) If you are not following the recommended version and use a svydesign object for the design you can still use weights to associate weights to your estimation. Beware that it might not give accurate estimations of the variance nor the uncertainty intervals.

.options.future

List of additional options for doFuture::%dofuture%().

...

Additional parameters for svrep::as_bootstrap_design().

Value

A pif_class() object containing the bootstrap simulations for the potential impact fraction, the average relative risk, and the average counterfactual if applicable.

Additional parallelization options

Faster computation occurs when doing parallelization which allows to use more cores in your machine. Parallelization utilizes the future::future() package. For paralelization to work you need to establish a plan (see future::plan() for more information). The most common way to create parallelization in your local machine is to do:

plan(multisession) #Start parallelization
pif(...)
plan(sequential) #Return to 'normal'

References

Chan CE, Zepeda-Tello R, Camacho-García-Formentí D, Cudhea F, Meza R, Rodrigues E, Spiegelman D, Barrientos-Gutierrez T, Zhou X (2023). “Nonparametric Estimation of the Potential Impact Fraction and Population Attributable Fraction with Individual-Level and Aggregated Data.” 2207.03597.

See also

Examples

# Use the ensanut dataset
data(ensanut)

# EXAMPLE 1
# Setup the survey design
options(survey.lonely.psu = "adjust")
design <- survey::svydesign(data = ensanut, ids = ~1, weights = ~weight, strata = ~strata)
rr <- function(X, theta) {
  exp(-2 +
    theta[1] * X[, "age"] + theta[2] * X[, "systolic_blood_pressure"] / 100)
}
cft <- function(X) {
  X[, "systolic_blood_pressure"] <- X[, "systolic_blood_pressure"] - 5
  return(X)
}
pif(design,
  theta = log(c(1.05, 1.38)), rr, cft,
  additional_theta_arguments = c(0.01, 0.03), n_bootstrap_samples = 10,
)
#> ── Potential Impact Fraction (PIF) ─────────────────────────────────────────────
#>     counterfactual   relative_risk potential_impact_fraction
#> 1 Counterfactual_1 Relative_Risk_1               0.014760072
#> 2 Counterfactual_1 Relative_Risk_1              -0.001760528
#> 3 Counterfactual_1 Relative_Risk_1               0.031280671
#>   average_relative_risk average_counterfactual           type
#> 1              5607.727               5509.783 point_estimate
#> 2            -15080.958             -14810.798     Lower 2.5%
#> 3             26296.412              25830.363    Upper 97.5%
#> ────────────────────────────────────────────────────────────────────────────────
#> • Number of bootstrap simulations: 10
#> ✖ A low number of bootstrap simulations will result in an unstable estimate.
#> • Use `as.data.frame` to access values.
#> • Use `summary` to save list of main results.

# EXAMPLE 2
# Now do the same but using a replicate design
rep_design <- svrep::as_bootstrap_design(design, replicates = 10)
pif(rep_design,
  theta = log(c(1.05, 1.38)), rr, cft,
  additional_theta_arguments = c(0.01, 0.03)
)
#> ── Potential Impact Fraction (PIF) ─────────────────────────────────────────────
#>     counterfactual   relative_risk potential_impact_fraction
#> 1 Counterfactual_1 Relative_Risk_1               0.016760903
#> 2 Counterfactual_1 Relative_Risk_1              -0.001709183
#> 3 Counterfactual_1 Relative_Risk_1               0.035230988
#>   average_relative_risk average_counterfactual           type
#> 1               47078.8               46254.13 point_estimate
#> 2             -171436.2             -168512.93     Lower 2.5%
#> 3              265593.8              261021.20    Upper 97.5%
#> ────────────────────────────────────────────────────────────────────────────────
#> • Number of bootstrap simulations: 10
#> ✖ A low number of bootstrap simulations will result in an unstable estimate.
#> • Use `as.data.frame` to access values.
#> • Use `summary` to save list of main results.

# EXAMPLE 3
# Calculate two different relative risks
rr <- list(
 function(X, theta) {exp(theta[1] * X[, "systolic_blood_pressure"] / 100)},
 function(X, theta) {exp(theta[2] * X[, "systolic_blood_pressure"] / 100 + theta[3]* X[,"age"])}
)

# Calculate three counterfactual scenarios of SBP reduction from 1 to 3 mmhg
cft <- list(
function(X){X[, "systolic_blood_pressure"]  <- X[, "systolic_blood_pressure"]  - 1; return(X)},
function(X){X[, "systolic_blood_pressure"]  <- X[, "systolic_blood_pressure"]  - 2; return(X)},
function(X){X[, "systolic_blood_pressure"]  <- X[, "systolic_blood_pressure"]  - 3; return(X)}
)

pif(design,
  theta = log(c(1.05, 1.38, 1.21)), rr, cft,
  additional_theta_arguments = c(0.01, 0.03, 0.025),
  n_bootstrap_samples = 10,
)
#> ── Potential Impact Fraction (PIF) ─────────────────────────────────────────────
#>      counterfactual   relative_risk potential_impact_fraction
#> 1  Counterfactual_1 Relative_Risk_1              0.0003597566
#> 2  Counterfactual_2 Relative_Risk_1              0.0007182515
#> 3  Counterfactual_3 Relative_Risk_1              0.0010754862
#> 4  Counterfactual_1 Relative_Risk_2              0.0031302785
#> 5  Counterfactual_2 Relative_Risk_2              0.0062467875
#> 6  Counterfactual_3 Relative_Risk_2              0.0093495944
#> 7  Counterfactual_1 Relative_Risk_1             -0.0021394858
#> 8  Counterfactual_1 Relative_Risk_1              0.0028589991
#> 9  Counterfactual_2 Relative_Risk_1             -0.0042777203
#> 10 Counterfactual_2 Relative_Risk_1              0.0057142233
#> 11 Counterfactual_3 Relative_Risk_1             -0.0064147075
#> 12 Counterfactual_3 Relative_Risk_1              0.0085656799
#> 13 Counterfactual_1 Relative_Risk_2             -0.0015499519
#> 14 Counterfactual_1 Relative_Risk_2              0.0078105089
#> 15 Counterfactual_2 Relative_Risk_2             -0.0030848133
#> 16 Counterfactual_2 Relative_Risk_2              0.0155783882
#> 17 Counterfactual_3 Relative_Risk_2             -0.0046046899
#> 18 Counterfactual_3 Relative_Risk_2              0.0233038787
#>    average_relative_risk average_counterfactual           type
#> 1           1.052909e+00           1.052386e+00 point_estimate
#> 2           1.052909e+00           1.051865e+00 point_estimate
#> 3           1.052909e+00           1.051345e+00 point_estimate
#> 4           5.892535e+18           5.885971e+18 point_estimate
#> 5           5.892535e+18           5.879414e+18 point_estimate
#> 6           5.892535e+18           5.872865e+18 point_estimate
#> 7           7.350943e-01           7.374154e-01     Lower 2.5%
#> 8           1.370724e+00           1.367358e+00    Upper 97.5%
#> 9           7.350943e-01           7.397347e-01     Lower 2.5%
#> 10          1.370724e+00           1.363996e+00    Upper 97.5%
#> 11          7.350943e-01           7.420521e-01     Lower 2.5%
#> 12          1.370724e+00           1.360639e+00    Upper 97.5%
#> 13         -3.556765e+19          -3.552826e+19     Lower 2.5%
#> 14          4.735272e+19           4.730020e+19    Upper 97.5%
#> 15         -3.556765e+19          -3.548891e+19     Lower 2.5%
#> 16          4.735272e+19           4.724774e+19    Upper 97.5%
#> 17         -3.556765e+19          -3.544961e+19     Lower 2.5%
#> 18          4.735272e+19           4.719534e+19    Upper 97.5%
#> ────────────────────────────────────────────────────────────────────────────────
#> • Number of bootstrap simulations: 10
#> ✖ A low number of bootstrap simulations will result in an unstable estimate.
#> • Use `as.data.frame` to access values.
#> • Use `summary` to save list of main results.